Create README.md
Thanks to the changing and developing technological structures, the supply-demand balance for cash in the human ecosystem has jumped to a different lane with the emergence of blockchain technologies and digital money system.
With credit cards, contactless payments, cryptocurrencies, and now digital money studies, it is clear that the thousands of years of humankind's relationship with cash will no longer be the same as before.
That's why; FABİ Token will be able to replace cash, meal vouchers and tickets, which were once used as a means of payment in the food and beverage industry in the developing digital world.
FABI Token; it is not just built to meet individual demands. FABI Token will bring convenience to the Food and Beverage Industry for all corporate companies and businesses operating in different sectors all over the world.